+++ interpals_api/__init__.py
@@ -1,3 +1,3 @@
-from .lib.session import Session, SessionAsync
-from .api import Api, ApiAsync
-import parsers
+# from .lib.session import Session, SessionAsync
+# from .api import Api, ApiAsync
+# import parsers
