+++ interpals_api/job/job_configurations.py
@@ -1,42 +1,12 @@
-from enum import Enum
-from typing import List, Optional
 from fastapi import HTTPException
-from pydantic import BaseModel
-import parsers
+from .parsers import parse_and_validate_search_options, parse_cron_from_date
 from .validate import validate_days
-from lib.constants import ContinentCode, CountryCode, Genders, SortOptions
-from store.store import redis_client
-from lib.errors import ExistingKeyException
-
+from ..store.store import redis_client
+from ..lib.errors import ExistingKeyException
+from .models import JobConfigRequest, JobType
 import json
 
-class JobType(Enum):
-    SEARCH = "search"
-    CHAT = "chat"
 
-class JobConfigRequest(BaseModel): 
-    job_name: str
-    min: int
-    hour: int
-    days: List[str]
-    type: JobType
-    data: Optional[object] = None
-
-
-class SearchOptions(BaseModel):
-    age1: Optional[str] = "16"
-    age2: Optional[str] = "110"
-    sex: Optional[List[Genders]] = ["male", "female"]
-    continents: Optional[List[ContinentCode]] = ["AF", "AS", "EU", "NA", "OC", "SA"]
-    countries: Optional[List[CountryCode]] = ["---"]
-    keywords: Optional[str] = ""
-    online: Optional[bool] = False
-    photo: Optional[bool] = False
-    city: Optional[str] = None
-    cityName: Optional[str] = None
-    sort: SortOptions = SortOptions.NEWEST_FIRST
-    limit: Optional[int] = 1000
-    timeout: Optional[float] = 0.0
 
 
 REDIS_JOB_BASE_KEY = "config:job"
@@ -50,10 +20,10 @@ async def add_cron_job(job: JobConfigRequest):
         #todo: add other cron job validations - jobs must be spaced at least an hour apart, not more than a specified number daily
         days = validate_days(job.days)
         if job.type == JobType.SEARCH:
-            job.data = parsers.parse_and_validate_search_options(job.data)
+            job.data = parse_and_validate_search_options(job.data)
 
         job_dict = {
-            "cron_time": parsers.parse_cron_from_date(job.min, job.hour, days),
+            "cron_time": parse_cron_from_date(job.min, job.hour, days),
             "job_type": str(job.type),
             "data": job.data,
         }
